requirements.txt
mysql-connector-python
python-dotenv
websocket-client==0.56